EXPERIMENTAL - Introduce Select mode
This behaves exactly like Eclipse's selection: any movement clears the
selection and moves back to plain insert, any alphanumeric characters
replace the selection with that character.

Extending with Shift+Movement works as usual but starting Select mode is
not yet implemented. For now it can only be entered when 'visualother'
is set and a selection is started in Insert mode.